Blog List Expansion, step 11 done

+++ part4/blogList/controllers/blogs.js
@@ -1,16 +1,16 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const { userExtractor } = require("../utils/middleware")
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
-
-  const user = await User.findOne({})
+  const user = await User.findById(request.user.id)
 
   const blog = new Blog({
     title: body.title,
@@ -27,9 +27,23 @@ blogsRouter.post('/', async (request, response) => {
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+  const user = await User.findById(request.user.id)
+  if (!user) {
+    return response.status(404).json({ error: 'user not found' })
+  }
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    return response.status(204).end()
+  }
+
+  return response.status(403).json({ error: 'only creator can delete post' })
+
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
